refactored out AppendBlock interface
